Добавил контроллер и сервис обработки заказов

+++ orders_service/src/main/java/com/example/orders_service/dto/DishToBeOrderedDto.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class DishToBeOrderedDto {
     long id;
 
-    long count;
+    int count;
 }
